# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "QR-based Event Entry Management Platform API",
        "title": "Event Entry Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/analytics/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get dashboard analytics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.AnalyticsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.AuthResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.AuthResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/coordinators": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coordinators"
                ],
                "summary": "List coordinators (Master only)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.CoordinatorResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coordinators"
                ],
                "summary": "Create coordinator (Master only)",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.CreateCoordinatorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/coordinators/{id}/disable": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Coordinators"
                ],
                "summary": "Disable coordinator (Master only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Coordinator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/coordinators/{id}/reset-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Coordinators"
                ],
                "summary": "Reset coordinator password (Master only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Coordinator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.ResetPasswordResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/guests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "List guests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.PaginatedGuestsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Create guest",
                "parameters": [
                    {
                        "description": "Guest data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.CreateGuestRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.GuestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/guests/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Import guests from CSV/XLSX",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV or XLSX file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.ImportResult"
                        }
                    }
                }
            }
        },
        "/api/v1/guests/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Search guests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.PaginatedGuestsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/guests/verify": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Manual verification search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.GuestSearchResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/guests/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Get guest by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.GuestResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Update guest (Master only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Guest data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.UpdateGuestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.GuestResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Delete guest (Master only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/insights": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Insights"
                ],
                "summary": "Get insights",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.InsightsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/reports/export/csv": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Export guests CSV report",
                "responses": {}
            }
        },
        "/api/v1/reports/export/pdf": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Export PDF report",
                "responses": {}
            }
        },
        "/api/v1/scan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan"
                ],
                "summary": "Scan QR code for entry",
                "parameters": [
                    {
                        "description": "QR token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.ScanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.ScanResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_sameetpatro_go-qr-auth_internal_dto.AnalyticsResponse": {
            "type": "object",
            "properties": {
                "coordinator_entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.CoordinatorEntryCount"
                    }
                },
                "hourly_entry_count": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.HourlyEntryCount"
                    }
                },
                "master_entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.CoordinatorEntryCount"
                    }
                },
                "overview": {
                    "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.AnalyticsOverview"
                }
            }
        },
        "github_com_sameetpatro_go-qr-auth_internal_dto.AuthResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_sameetpatro_go-qr-auth_internal_dto.UserResponse"
                }
            }
        },
...
        "github_com_sameetpatro_go-qr-auth_internal_dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}